move graphql ID type to shared/graphql
