+++ apps/client/src/shared/graphql/index.ts
@@ -2,3 +2,5 @@
 export { type Cursor, GraphqlConnection } from "./connection";
 /** @public */
 export { useConnection } from "./hook";
+/** @public */
+export type { ID } from "./types";
