+++ apps/client/src/shared/graphql/types.ts
@@ -1,0 +1,3 @@
+import type { Brand } from "@uzmoi/ut/types";
+
+export type ID = string & Brand<"GraphQL/ID">;
